event-trigger payment tracking

+++ contracts/Item.sol
@@ -3,13 +3,17 @@ pragma solidity ^0.6.0;
 import "./ItemManager.sol";
 
 contract Item {
-    uint public priceInWei;
-    uint public paidWei;
-    uint public index;
+    uint256 public priceInWei;
+    uint256 public paidWei;
+    uint256 public index;
 
     ItemManager parentContract;
 
-    constructor(ItemManager _parentContract, uint _priceInWei, uint _index) public {
+    constructor(
+        ItemManager _parentContract,
+        uint256 _priceInWei,
+        uint256 _index
+    ) public {
         priceInWei = _priceInWei;
         index = _index;
         parentContract = _parentContract;
@@ -19,12 +23,11 @@ contract Item {
         require(msg.value == priceInWei, "We don't support partial payments");
         require(paidWei == 0, "Item is already paid!");
         paidWei += msg.value;
-        (bool success, ) = address(parentContract).call{value:msg.value}(abi.encodeWithSignature("triggerPayment(uint256)", index));
+        (bool success, ) = address(parentContract).call{value: msg.value}(
+            abi.encodeWithSignature("triggerPayment(uint256)", index)
+        );
         require(success, "Delivery did not work");
     }
 
-    fallback () external {
-
-    }
-
+    fallback() external {}
 }
